1. 增加国网界面修改 2. 增加代理13762透传帧功能 3. 完善控制功能 4. 增加本地模块互换性功能 5. 增加加密抄读功能

+++ ClouSoft/common/LibDb/LibDbAPI.h
@@ -18,7 +18,7 @@
 
 
 TItemDesc* BinarySearchItem(TItemDesc* pItemDesc, WORD num, WORD wID);
-int BinarySearchIndex(TItemDesc* pItemDesc, DWORD num, WORD wID);
+int BinarySearchIndex(TItemDesc* pItemDesc, WORD num, WORD wID);
 bool InitItemDesc(TBankCtrl* pBankCtrl);
 int ReadItem(WORD wImg, WORD wPn, WORD wID, TItemAcess& rItemAcess, 
 			 DWORD dwStartTime, DWORD dwEndTime, TBankCtrl* pBankCtrl);
